Updated thread behaviour
Thread now handles both the connection and processing.

The main thread's job is to simply check the success of the thread
and update the success or error log.

This ensures that there is not a build up of open streams that go
beyond the specified max thread count due to the main thread
handling all the open streams created by the threads.

stream is explicitly closed before the thread returns